Agno/Multi Agent Playground/requirements.txt
agno
google-genai
python-dotenv
duckduckgo-search
pypdf
lancedb
tantivy
streamlit
pandas
yfinance
sqlalchemy
fastapi[standard]